+++ lib/exwiw/adapter/sqlite3_adapter.rb
@@ -74,8 +74,12 @@ module Exwiw
         end
         values = value_list.join(",\n")
 
-        column_names = table.columns.map(&:name).join(', ')
-        "INSERT INTO #{table_name} (#{column_names}) VALUES\n#{values};"
+        if table.rails_managed?
+          "INSERT INTO #{table_name} VALUES\n#{values};"
+        else
+          column_names = table.columns.map(&:name).join(', ')
+          "INSERT INTO #{table_name} (#{column_names}) VALUES\n#{values};"
+        end
       end
 
       def to_bulk_delete(select_query_ast, table)
@@ -128,7 +132,11 @@ module Exwiw
         raise NotImplementedError unless query_ast.is_a?(Exwiw::QueryAst::Select)
 
         sql = "SELECT "
-        sql += query_ast.columns.map { |col| compile_column_name(query_ast, col) }.join(', ')
+        sql += if query_ast.select_all
+                 "*"
+               else
+                 query_ast.columns.map { |col| compile_column_name(query_ast, col) }.join(', ')
+               end
         sql += " FROM #{query_ast.from_table_name}"
 
         query_ast.join_clauses.each do |join|