+++ lib/exwiw/table_config.rb
@@ -4,16 +4,44 @@ module Exwiw
   class TableConfig
     include Serdes
 
+    RAILS_MANAGED_SCHEMA_MIGRATIONS = "rails_managed_schema_migrations"
+    RAILS_MANAGED_INTERNAL_METADATA = "rails_managed_internal_metadata"
+    RAILS_MANAGED_TYPES = [
+      RAILS_MANAGED_SCHEMA_MIGRATIONS,
+      RAILS_MANAGED_INTERNAL_METADATA,
+    ].freeze
+
     attribute :name, String
-    attribute :primary_key, String
+    attribute :primary_key, optional(String), skip_serializing_if_nil: true
+    attribute :type, optional(String), skip_serializing_if_nil: true
+    attribute :comment, optional(String), skip_serializing_if_nil: true
     attribute :filter, optional(String), skip_serializing_if_nil: true
-    attribute :belongs_tos, array(BelongsTo)
-    attribute :columns, array(TableColumn)
+    attribute :belongs_tos, array(BelongsTo), default: []
+    attribute :columns, array(TableColumn), default: []
     attribute :bulk_insert_chunk_size, optional(Integer), skip_serializing_if_nil: true
     attribute :skip, Serdes::OptionalType.new(Serdes::ConcreteType.new(Boolean)), skip_serializing_if_nil: true
 
+    def self.from(hash)
+      config = super
+      config.send(:validate_after_load!)
+      config
+    end
+
     def self.from_symbol_keys(hash)
       from(JSON.parse(hash.to_json))
+    end
+
+    def rails_managed?
+      RAILS_MANAGED_TYPES.include?(type)
+    end
+
+    def to_hash
+      hash = super
+      if rails_managed?
+        hash.delete("belongs_tos")
+        hash.delete("columns")
+      end
+      hash
     end
 
     def column_names
@@ -74,6 +102,8 @@ module Exwiw
       TableConfig.new.tap do |merged_table|
         merged_table.name = name
         merged_table.primary_key = passed_table.primary_key
+        merged_table.type = passed_table.type
+        merged_table.comment = comment
         merged_table.filter = filter
         merged_table.belongs_tos = passed_table.belongs_tos
         merged_table.bulk_insert_chunk_size = passed_table.bulk_insert_chunk_size
@@ -90,6 +120,27 @@ module Exwiw
               passed_column
             end
           end
+      end
+    end
+
+    private def validate_after_load!
+      if rails_managed?
+        if primary_key
+          raise ArgumentError,
+                "Table '#{name}' has type=#{type}; primary_key must not be defined."
+        end
+        if !belongs_tos.empty?
+          raise ArgumentError,
+                "Table '#{name}' has type=#{type}; belongs_tos must not be defined."
+        end
+        if !columns.empty?
+          raise ArgumentError,
+                "Table '#{name}' has type=#{type}; columns must not be defined."
+        end
+      else
+        if primary_key.nil?
+          raise ArgumentError, "Table '#{name}' requires primary_key."
+        end
       end
     end